+++ tests/test_table_xls.py
@@ -74,3 +74,12 @@ class TestXLS(agate.AgateTestCase):
         self.assertColumnNames(table, [])
         self.assertColumnTypes(table, [])
         self.assertRows(table, [])
+
+    def test_numeric_column_name(self):
+        table = agate.Table.from_xls('examples/test_numeric_column_name.xls')
+
+        self.assertColumnNames(table, ('Country', '2013.0'))
+        self.assertColumnTypes(table, [agate.Text, agate.Number])
+        self.assertRows(table, [
+            ['Canada', 35160000]
+        ])