+++ tests/test_table_xlsx.py
@@ -63,3 +63,12 @@ class TestXLSX(agate.AgateTestCase):
         self.assertColumnNames(table, [])
         self.assertColumnTypes(table, [])
         self.assertRows(table, [])
+
+    def test_numeric_column_name(self):
+        table = agate.Table.from_xlsx('examples/test_numeric_column_name.xlsx')
+
+        self.assertColumnNames(table, ['Country', '2013'])
+        self.assertColumnTypes(table, [agate.Text, agate.Number])
+        self.assertRows(table, [
+            ['Canada', 35160000]
+        ])